Refactored upgrade function.
